+++ Source/ACE.Entity/Enum/ParticleType.cs
@@ -1,0 +1,20 @@
+﻿namespace ACE.Entity.Enum
+{
+    public enum ParticleType
+    {
+        Unknown = 0,
+        Still = 1,
+        LocalVelocity = 2,
+        ParabolicLVGA = 3,
+        ParabolicLVGAGR = 4,
+        Swarm = 5,
+        Explode = 6,
+        Implode = 7,
+        ParabolicLVLA = 8,
+        ParabolicLVLALR = 9,
+        ParabolicGVGA = 10,
+        ParabolicGVGAGR = 11,
+        GlobalVelocity = 12,
+        NumParticleType = 13,
+    }
+}